Create fragment to create new posts

+++ app/src/main/java/ch/epfl/swissteam/services/Post.java
@@ -9,7 +9,7 @@ package ch.epfl.swissteam.services;
 public class Post {
 
     private String title_, username_, body_;
-    private int timestamp_;
+    private long timestamp_;
 
     /**
      * Default constructor required for database.
@@ -26,7 +26,7 @@ public class Post {
      * @param body the body of the post
      * @param timestamp the timestamp at which the post was submitted
      */
-    public Post(String title, String username, String body, int timestamp) {
+    public Post(String title, String username, String body, long timestamp) {
         this.title_ = title;
         this.username_ = username;
         this.body_ = body;
@@ -45,7 +45,7 @@ public class Post {
         return body_;
     }
 
-    public int getTimestamp() {
+    public long getTimestamp() {
         return timestamp_;
     }
 }
